# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "Helpoo Api Documentation",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module, for <b>Helpoo</b>."
  },
  "servers": [
    {
      "url": "http://apidev.helpooapp.net/"
    },
    {
      "url": "https://apidev.helpooapp.net/"
    }
  ],
...
  "paths": {
...
    "/api/v2/carModels/manufacturer/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v2/inspect/createInspector": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "insuranceId": {
                    "example": "any"
                  },
                  "phoneNumbers": {
                    "example": "any"
                  },
                  "emails": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/inspect/addInspector": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "inspectionCompanyId": {
                    "example": "any"
                  },
                  "phoneNumbers": {
                    "example": "any"
                  },
                  "emails": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/inspect/getByInsurance/{insuranceId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "insuranceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/inspect/getByInspectionCompany/{inspectionCompanyId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "inspectionCompanyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/inspect/assignInspector": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accidentReportId": {
                    "example": "any"
                  },
                  "inspectorId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/manufacturers/create": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "create manufacturer",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "en_name": {
                  "type": "string",
                  "example": "en_name"
                },
                "ar_name": {
                  "type": "string",
                  "example": "ar_name"
                }
              },
              "required": [
                "en_name",
                "ar_name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/manufacturers/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "update manufacturer",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "en_name": {
                  "type": "string",
                  "example": "en_name"
                },
                "ar_name": {
                  "type": "string",
                  "example": "ar_name"
                }
              },
              "required": [
                "en_name",
                "ar_name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v2/paymob/getIframe": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "getIframe",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "amount": {
                  "type": "number",
                  "example": 1500
                },
                "packageId": {
                  "type": "number",
                  "example": 1
                },
                "serviceRequestId": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "amount",
                "packageId",
                "serviceRequestId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "packageId": {
                    "example": "any"
                  },
                  "serviceRequestId": {
                    "example": "any"
                  },
                  "amount": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/paymob/callback": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "success",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pending",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "txn_response_code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}